// apps/server/src/modules/common-cartridge/common-cartridge-client/lesson-client/lessons-api-client/models/component-nexboard-props-impl.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ComponentNexboardPropsImpl
 */
export interface ComponentNexboardPropsImpl {
    /**
     * board of a Nexboard component
     * @type {string}
     * @memberof ComponentNexboardPropsImpl
     */
    'board': string;
    /**
     * description of a Nexboard component
     * @type {string}
     * @memberof ComponentNexboardPropsImpl
     */
    'description': string;
    /**
     * title of a Nexboard component
     * @type {string}
     * @memberof ComponentNexboardPropsImpl
     */
    'title': string;
    /**
     * url of a Nexboard component
     * @type {string}
     * @memberof ComponentNexboardPropsImpl
     */
    'url': string;
}